chore: move to utils.
