+++ routes/_middleware.ts
@@ -1,10 +1,13 @@
 import { MiddlewareHandler } from "$fresh/server.ts";
-import { createServerClient } from "../components/supabase.ts";
+import { createServerClient } from "../utils/supabase.ts";
 
 export const handler: MiddlewareHandler[] = [
   async function getSupabaseSession(req, ctx) {
     const { pathname } = new URL(req.url);
-    if (pathname.includes("_frsh") || pathname.includes(".ico")) {
+    if (
+      pathname.includes("api/auth/signout") || pathname.includes("_frsh") ||
+      pathname.includes(".ico")
+    ) {
       return ctx.next();
     }
     const headers = new Headers();
