+++ routes/signup.tsx
@@ -1,5 +1,20 @@
 import { Head } from "$fresh/runtime.ts";
+import type { Handlers } from "$fresh/server.ts";
 import SignUpOrIn from "../islands/SignUpOrIn.tsx";
+
+export const handler: Handlers = {
+  GET(_req, ctx) {
+    if (ctx.state.session) {
+      const headers = new Headers();
+      headers.set("location", "/");
+      return new Response(null, {
+        status: 303,
+        headers,
+      });
+    }
+    return ctx.render!();
+  },
+};
 
 export default function SignUpPage() {
   return (
